Vertx client

+++ WerkCore/src/main/java/org/werk/processing/readonly/ReadOnlyJob.java
@@ -8,5 +8,5 @@ import org.werk.data.StepPOJO;
 public interface ReadOnlyJob<J> extends JobPOJO<J> {
 	Collection<StepPOJO> getProcessingHistory() throws Exception;
 	Collection<StepPOJO> getFilteredHistory(String stepType) throws Exception;
-	StepPOJO getStep(long stepNumber) throws Exception;
+	StepPOJO getStep(int stepNumber) throws Exception;
 }
